pyproject.toml
[project]
name = "headshot-generator"
version = "0.1.0"
description = "Interactive headshot generator using OpenCV and Streamlit"
readme = "README.md"
authors = [
    { name = "Michael Booth", email = "michael@databooth.com.au" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "streamlit>=1.50.0",
    "toml>=0.10.2",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[tool.uv]
package = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
    "src",
]
